// web/src/api/generated/models/banner-list-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 轮播图列表响应
 */
export interface BannerListResponse {
    /**
     * 轮播图ID
     */
    'id'?: number;
    /**
     * 标题
     */
    'title'?: string;
    /**
     * 图片URL
     */
    'imageUrl'?: string;
    /**
     * 跳转类型: 0-无跳转, 1-内部路由, 2-外部链接
     */
    'linkType'?: number;
    /**
     * 跳转URL/路由
     */
    'linkUrl'?: string;
}